Create WriteTxnToContract
With this contract transaction interaction we can send a transaction from account 1 to account 2 via the smart contract being referenced. Important to note the transaction via web3 is not the same as the smart contract. The smart contract can interact with transactions  and is the manager for the transaction from account 1 to account 2. So this works in tandem with our smart contract which was manually deployed via app.js in the ContractDeployment file.