Create Web3 Event Subscription
This is an example of event subscription. This will pump out all the data and important variables associated with the transaction such as block number, transaction index, blockhash, logIndex, etc which you can interact with using another smart contract or create a dapp/ui that pulls directly from this web3 data.